ActivityIndicatorのグルグル状態をboolValueとしてもつActivityIndicatorCellを作成
Frame/Grid/StackPanel/ScrollCell のコンストラクタで、View引数を省略したときに、内部でビューを作成するように修正
mmにすると、floatValue が、NSNumber.floatValueと型が異なってエラーになるケースがあったので、floatに変更。
WPLObserverクラスを追加

+++ layouter/wpl/observable/WPLObservableDef.h
@@ -21,7 +21,7 @@
      */
     @property (nonatomic,readonly) id value;
     @property (nonatomic,readonly) NSString* stringValue;
-    @property (nonatomic,readonly) CGFloat floatValue;
+    @property (nonatomic,readonly) float floatValue;
     @property (nonatomic,readonly) bool boolValue;
     @property (nonatomic,readonly) NSInteger integerValue;
     @property (nonatomic,readonly) int intValue;
@@ -77,7 +77,7 @@
      */
     @property (nonatomic) id value;
     @property (nonatomic) NSString* stringValue;
-    @property (nonatomic) CGFloat floatValue;
+    @property (nonatomic) float floatValue;
     @property (nonatomic) bool boolValue;
     @property (nonatomic) int intValue;
     @property (nonatomic) NSInteger integerValue;
